output smt2

+++ test.c
@@ -2,7 +2,7 @@
     #include "seahorn/seahorn.h"
     
     int main(void){
-      float x = 4.0;
+      int x = 4;
       int y = 3;
       int n = unknown();
       while (y<n){
